Move TransactionStatus enum
