+++ src/main/scala/org/constellation/storage/transactions/TransactionStatus.scala
@@ -1,0 +1,6 @@
+package org.constellation.storage.transactions
+
+object TransactionStatus extends Enumeration {
+  type TransactionStatus = Value
+  val Pending, Arbitrary, InConsensus, Accepted, Unknown = Value
+}
